% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vos_install.R
\name{has_virtuoso}
\alias{has_virtuoso}
\title{check for Virtuoso}
\usage{
has_virtuoso()
}
\value{
logical indicating if virtuoso-t binary was found or now.
}
\description{
test if the system has a virtuoso installation on the path
}
\examples{
has_virtuoso()
}
