% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_.R
\name{sparql_op}
\alias{sparql_op}
\title{sparql_select("name", "license", "author.familyName")}
\usage{
sparql_op(select = character(), where = character(),
  from = character())
}
\description{
rename a variable
sparql_select(package = "name", "license", coauthor = "author.familyName") %>%
sparql_filter(author.familyName == "Boettiger", author.givenName == "Carl") %>%
sparql_build()
}
