% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_.R
\name{sparql_select}
\alias{sparql_select}
\title{sparql_select("name", "license", "author.familyName")}
\usage{
sparql_select(..., prefix = NULL, na.rm = TRUE)
}
\description{
sparql_select("name", "license", "author.familyName")
}
