% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtuoso-package.R
\docType{package}
\name{virtuoso-package}
\alias{virtuoso}
\alias{virtuoso-package}
\title{virtuoso: An R Interface to Virtuoso Using ODBC}
\description{
Virtuoso is a high-performance "universal server," which can act
as both a relational database (supporting standard SQL queries),
and an Resource Description Framework (RDF) triplestore, supporting
SPARQL queries and semantic reasoning. The \code{virtuoso} R package provides
R users with a DBI-compatible connection to the Virtuoso database.
The package also provides helper routines to install, launch, and manage
a Virtuoso server locally on Mac, Windows and Linux platforms using
the standard interactive installers from the R command-line.  By
automatically handling these setup steps, the package can make Virtuoso
considerably faster and easier for a most users to deploy in a local
environment. While this can be used as a normal \code{dplyr} backend, Virtuoso
excels when used as a RDF triplestore.  Managing the bulk import of triples
from common serializations with a single intuitive command is another key
feature of the \code{virtuoso} R package.  Bulk import performance can be tens to
hundreds of times faster than the comparable imports using existing R tools,
including \code{rdflib} and \code{redland} packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropensci/virtuoso}
  \item Report bugs at \url{https://github.com/ropensci/virtuoso/issues}
}

}
\author{
\strong{Maintainer}: Carl Boettiger \email{cboettig@gmail.com} (0000-0002-1642-628X) [copyright holder]

Other contributors:
\itemize{
  \item Bryce Mecum \email{brycemecum@gmail.com} (0000-0002-0381-3766) [contributor]
}

}
\keyword{internal}
