% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vos_configure.R
\name{vos_configure}
\alias{vos_configure}
\title{Configure Virtuoso Server ini file}
\usage{
vos_configure(DirsAllowed = ".", gigs_ram = 2,
  template = find_virtuoso_ini(),
  db_dir = rappdirs::user_log_dir("Virtuoso"))
}
\arguments{
\item{DirsAllowed}{Paths (relative or absolute) to directories from which
Virtoso should have read and write access (e.g. for bulk uploading). Should
be specificied as a single comma-separated string.  Default is the current
working directory.}

\item{gigs_ram}{Indicate approximately the maximum GB of memory Virtuoso can
have access to.  (Used to set NumberOfBuffers & MaxDirtyBuffers in config.)}

\item{template}{Location of an existing virtuoso.ini file which will be used
as a template. By default, `vos_configure()` will attempt to locate the
appropriate template for your system.}

\item{db_dir}{location where `virtuoso.ini` file should be written.  Other
Virtuoso database log files will also be written here.}
}
\value{
Writes the requested `virtuoso.ini` file to the db_dir specified
and returns the path to this file.
}
\description{
Virtuoso Server configuration is determined by a virtuoso.ini file when
server starts. This file includes both system-specific information from
your install (location of server files, addons, etc) and user-configurable
parameters. This helper function provides a way to create and modify an
appropriate `virtuoso.ini` file.
}
