% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtuoso_paths.R
\name{vos_destroy_all}
\alias{vos_destroy_all}
\title{Destroy all Virtuoso's directories}
\usage{
vos_destroy_all(force = FALSE)
}
\arguments{
\item{force}{should permissions be changed (if possible) to allow deletion?}
}
\value{
TRUE if entirely successful in removing all files,
 FALSE otherwise (invisibly).
}
\description{
for a clean reset of the system that purges all
data files, config files, cache and log files created
by virtuoso R package.
}
