% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vos_import.R
\name{vos_import}
\alias{vos_import}
\title{vos_import}
\usage{
vos_import(con, files = NULL, wd = ".", glob = "*",
  graph = "rdflib")
}
\arguments{
\item{con}{a ODBC connection to Virtuoso, from [`vos_connect()`]}

\item{files}{paths to files to be imported}

\item{wd}{The directory from which we can search the files.
NOTE: This directory must match or be in the directory from which you ran vos_start().
This is the default behavior and probably best not altered.}

\item{glob}{A wildcard aka globbing pattern (e.g. "*.nq").}

\item{graph}{Name (technically URI) for a graph in the database.  Can leave as default.
If a graph is already specified by the import file (e.g. in nquads), that will be used
instead.}
}
\description{
vos_import
}
\references{
http://vos.openlinksw.com/owiki/wiki/VOS/VirtBulkRDFLoader
}
