% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vos_install.R
\name{vos_install}
\alias{vos_install}
\title{Helper method for installing Virtuoso Server on Mac OSX}
\usage{
vos_install(use_brew = FALSE, prompt = is_interactive())
}
\arguments{
\item{use_brew}{Should we use homebrew to install? (Mac OSX only)}

\item{prompt}{Should we prompt user for interactive installation?}
}
\description{
Helper method for installing Virtuoso Server on Mac OSX
}
