% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vos_kill.R
\name{vos_kill}
\alias{vos_kill}
\title{Stop (kill) the Virtuoso server}
\usage{
vos_kill(p = NA)
}
\arguments{
\item{p}{a process object, returned by
[`vos_process()`] or  [`vos_start()`]}
}
\description{
Kill ends the process started by [`vos_start()`]
}
