% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtuoso.R
\name{vos_list_graphs}
\alias{vos_list_graphs}
\title{List graphs}
\usage{
vos_list_graphs(con)
}
\arguments{
\item{con}{a ODBC connection to Virtuoso, from \code{\link[=vos_connect]{vos_connect()}}}
}
\description{
List graphs
}
\examples{
vos_status()
\donttest{
if(has_virtuoso()){
vos_start()
con <- vos_connect()
vos_list_graphs(con)

}}
}
