% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vos_log.R
\name{vos_log}
\alias{vos_log}
\title{Query the server logs}
\usage{
vos_log(p = NA, collapse = NULL, just_errors = FALSE)
}
\arguments{
\item{p}{a process object, returned by
[`vos_process()`] or  [`vos_start()`]}

\item{collapse}{an optional character string to separate the
lines in a single character string.}

\item{just_errors}{logical, default FALSE. Set to true to return
just the lines that contain the term "error", which can be useful
in debugging or validating bulk imports.}
}
\description{
Query the server logs
}
