% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odbc.R
\name{vos_odbcinst}
\alias{vos_odbcinst}
\title{Configure the ODBC Driver for Virtuoso}
\usage{
vos_odbcinst(system_odbcinst = find_odbcinst(),
  local_odbcinst = odbcinst_path())
}
\arguments{
\item{system_odbcinst}{Path to the system odbcinst.ini file. (Does not
require write access.) Default will attempt to find the file for your system.}

\item{local_odbcinst}{Path to the local odbcinst we should use.}
}
\description{
Configure the ODBC Driver for Virtuoso
}
\details{
This function is called automatically by [`vos_install()`] and thus
does not usually need to be called by the user.  Users can also manually configure
ODBC as outlined in <https://github.com/r-dbi/odbc#dsn-configuration-files>.
This is merely a convenience function automating that process on most systems.

ODBC Uses a odbcinst.ini file to point ODBC at the library required
to drive any given database.  This function helps us automatically
locate the driver library on different operating systems and configure
the odbcinst appropriately for each OS.
}
