% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vos_process.R
\name{vos_process}
\alias{vos_process}
\title{Return a handle to an existing Virtuoso Process}
\usage{
vos_process(p = NA)
}
\arguments{
\item{p}{a process object, returned by
[`vos_process()`] or  [`vos_start()`]}
}
\description{
Return a handle to an existing Virtuoso Process
}
