% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vos_query.R
\name{vos_query}
\alias{vos_query}
\title{Run SPARQL query}
\usage{
vos_query(con, query)
}
\arguments{
\item{con}{a ODBC connection to Virtuoso, from [`vos_connect()`]}

\item{query}{a SPARQL query statement}
}
\description{
Run SPARQL query
}
