% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtuoso_paths.R
\name{vos_set_paths}
\alias{vos_set_paths}
\title{set Virtuoso paths}
\usage{
vos_set_paths(db_dir = vos_db(), config_dir = vos_config(),
  cache_dir = vos_cache(), log_dir = vos_logdir(),
  home = virtuoso_home())
}
\value{
A logical vector, with elements being true
if setting the corresponding variable succeeded
(invisibly).
}
\description{
Set the location of Virtuoso database, configure files,
cache, and logs to your preferred location.  Set home
to the location of your Virtuoso installation.
}
\examples{
vos_set_paths()

}
