% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odbc.R
\name{vos_start}
\alias{vos_start}
\title{Start a local Virtuoso Server}
\usage{
vos_start(ini = NULL)
}
\arguments{
\item{ini}{path to a virtuoso.ini configuration file. If not
provided, function will attempt to determine the location of the
default configuration file.}
}
\description{
Start a local Virtuoso Server
}
