% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vos_start.R
\name{vos_status}
\alias{vos_status}
\title{Query the server status}
\usage{
vos_status(p = NA)
}
\arguments{
\item{p}{a process object, returned by
[`vos_process()`] or  [`vos_start()`]}
}
\description{
Query the server status
}
