% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vos_uninstall.R
\name{vos_uninstall}
\alias{vos_uninstall}
\title{Uninstall Virtuoso}
\usage{
vos_uninstall()
}
\description{
Automatic uninstaller for Mac OSX and Windows clients.
}
\examples{
\dontrun{
vos_uninstall()
}

}
