% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vos_install.R
\name{vos_uninstall_windows}
\alias{vos_uninstall_windows}
\title{Uninstall Virutoso using Windows Uninstaller}
\usage{
vos_uninstall_windows(vos_home = virtuoso_home_windows())
}
\arguments{
\item{vos_home}{Home directory for Virtuoso}
}
\description{
Uninstall Virutoso using Windows Uninstaller
}
