% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_nquads.R
\name{write_nquads}
\alias{write_nquads}
\title{write object out as nquads}
\usage{
write_nquads(x, file, prefix = NULL, ...)
}
\arguments{
\item{x}{an object that can be represented as nquads}

\item{file}{output filename}

\item{prefix}{URI prefix that should be used to identify column names.
Can be of the form "http://schema.org/", or "iris:".}

\item{...}{additional parameters, see examples}
}
\description{
write object out as nquads
}
\examples{
tmp <- tempfile(fileext = ".nq")
library(datasets)
write_nquads(iris, tmp)
}
