% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtuoso.R
\name{vos_clear_graph}
\alias{vos_clear_graph}
\title{Clear all triples from a graph}
\usage{
vos_clear_graph(con, graph = "rdflib")
}
\arguments{
\item{con}{a ODBC connection to Virtuoso, from [`vos_connect()`]}

\item{graph}{Name (technically URI) for a graph in the database.
Can leave as default. If a graph is already specified by the
import file (e.g. in nquads), that will be used instead.}
}
\description{
Clear all triples from a graph
}
\details{
NOTE: after clearing a graph, re-running the bulk importer may not
re-import triples, at least until virtuoso-server is restarted.
}
