% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odbc.R
\name{vos_connect}
\alias{vos_connect}
\title{Connect to a Virtuoso Server over ODBC}
\usage{
vos_connect(driver = "Virtuoso", uid = "dba", pwd = "dba",
  host = "localhost", port = "1111")
}
\arguments{
\item{driver}{Name of the Driver line in the ODBC configuration}

\item{uid}{User id. Defaults to "dba"}

\item{pwd}{Password. Defaults to "dba"}

\item{host}{IP address of the Virtuoso Server}

\item{port}{Port used by Virtuoso. Defaults to
the Virtuoso standard port, 1111}
}
\description{
Connect to a Virtuoso Server over ODBC
}
