% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vos_destroy_all.R
\name{vos_delete_db}
\alias{vos_delete_db}
\title{Delete Virtuoso Database}
\usage{
vos_delete_db(ask = is_interactive(), db_dir = vos_db())
}
\arguments{
\item{ask}{ask before deleting?}

\item{db_dir}{location of the directory to delete}
}
\description{
delete the entire Virtuoso database for a fresh start.
}
\examples{

\dontshow{
virtuoso:::vos_test_paths()
}
vos_delete_db()

}
