% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vos_destroy_all.R
\name{vos_destroy_all}
\alias{vos_destroy_all}
\title{Destroy all Virtuoso's directories}
\usage{
vos_destroy_all(force = FALSE)
}
\arguments{
\item{force}{should permissions be changed (if possible) to allow deletion?}
}
\value{
\link{TRUE} if entirely successful in removing all files,
\link{FALSE} otherwise (invisibly).
}
\description{
Provides a clean reset of the system that purges all
data files, config files, cache and log files created
by virtuoso R package. This does not uninstall Virtuoso software
itself, see \code{\link[=vos_uninstall]{vos_uninstall()}} to uninstall.
}
\examples{

\dontshow{
virtuoso:::vos_test_paths()
}
vos_destroy_all()

}
