% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vos_kill.R
\name{vos_kill}
\alias{vos_kill}
\title{Stop (kill) the Virtuoso server}
\usage{
vos_kill(p = NA)
}
\arguments{
\item{p}{a process object, returned by
\code{\link[=vos_process]{vos_process()}} or  \code{\link[=vos_start]{vos_start()}}.
(will be restored from cache if not provided)}
}
\description{
Kill ends the process started by \code{\link[=vos_start]{vos_start()}}
}
\details{
vos_kill simply shuts down the local Virtuoso server,
it does not remove any data stored in the database system.
\code{\link[=vos_kill]{vos_kill()}} terminates the process, removing the
process id from the process table.
}
\examples{
\donttest{
if(has_virtuoso()){

  vos_start()
  vos_kill()

  }
}
}
\seealso{
\code{\link[=vos_start]{vos_start()}}
}
