% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vos_log.R
\name{vos_log}
\alias{vos_log}
\title{Query the server logs}
\usage{
vos_log(p = NA, collapse = NULL, just_errors = FALSE)
}
\arguments{
\item{p}{a process object, returned by
\code{\link[=vos_process]{vos_process()}} or  \code{\link[=vos_start]{vos_start()}}.
(will be restored from cache if not provided)}

\item{collapse}{an optional character string to separate the
lines in a single character string.}

\item{just_errors}{logical, default \link{FALSE}. Set to \link{TRUE} to return
just the lines that contain the term "error", which can be useful
in debugging or validating bulk imports.}
}
\value{
Virtuoso logs as a character vector.
}
\description{
Query the server logs
}
\examples{
if(has_virtuoso())
  vos_log()

}
\seealso{
\code{\link[=vos_start]{vos_start()}}
}
