% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtuoso.R
\name{vos_query}
\alias{vos_query}
\title{Run SPARQL query}
\usage{
vos_query(con, query, graph = "rdflib")
}
\arguments{
\item{con}{a ODBC connection to Virtuoso, from [`vos_connect()`]}

\item{query}{a SPARQL query statement}

\item{graph}{Name (technically URI) for a graph in the database.
Can leave as default. If a graph is already specified by the
import file (e.g. in nquads), that will be used instead.}
}
\description{
Run SPARQL query
}
