% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtuoso_paths.R
\name{vos_set_paths}
\alias{vos_set_paths}
\title{set Virtuoso paths}
\usage{
vos_set_paths(
  db_dir = vos_db(),
  config_dir = vos_config(),
  cache_dir = vos_cache(),
  log_dir = vos_logdir(),
  home = virtuoso_home()
)
}
\arguments{
\item{db_dir}{Location of data in the Virtuoso (tables, triplestore)}

\item{config_dir}{Location of configuration files for Virtuoso}

\item{cache_dir}{Location of cache for bulk importing}

\item{log_dir}{Location of Virutoso Server logs}

\item{home}{Location of the Virtuoso installation}
}
\value{
A logical vector, with elements being true
if setting the corresponding variable succeeded
(invisibly).
}
\description{
Set the location of Virtuoso database, configure files,
cache, and logs to your preferred location.  Set home
to the location of your Virtuoso installation.
}
\examples{
if(has_virtuoso())
  vos_set_paths()

}
