% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vos_start.R
\name{vos_start}
\alias{vos_start}
\title{Start a Virtuoso Server}
\usage{
vos_start(ini = NULL, wait = 30)
}
\arguments{
\item{ini}{path to a virtuoso.ini configuration file. If not
provided, function will attempt to determine the location of the
default configuration file.}

\item{wait}{number of seconds to wait for server to come online}
}
\description{
Start a Virtuoso Server
}
